role adjusted, persist-db added, etc.

+++ src/main/java/net/joedoe/mvcsech2/controllers/ProductController.java
@@ -26,7 +26,7 @@ public class ProductController {
     }
 
     @GetMapping("/product/{id}")
-    public String getProductById(@PathVariable Integer id, Model model) {
+    public String getProductById(@PathVariable Long id, Model model) {
         model.addAttribute("product", productService.getById(id));
         return "product";
     }
