+++ src/main/java/net/joedoe/mvcsech2/repositories/IProductRepository.java
@@ -3,5 +3,5 @@ package net.joedoe.mvcsech2.repositories;
 import net.joedoe.mvcsech2.domains.Product;
 import org.springframework.data.repository.CrudRepository;
 
-public interface IProductRepository extends CrudRepository<Product, Integer> {
+public interface IProductRepository extends CrudRepository<Product, Long> {
 }
