+++ src/main/java/net/joedoe/mvcsech2/repositories/IUserRepository.java
@@ -3,6 +3,6 @@ package net.joedoe.mvcsech2.repositories;
 import net.joedoe.mvcsech2.domains.User;
 import org.springframework.data.repository.CrudRepository;
 
-public interface IUserRepository extends CrudRepository<User, Integer> {
+public interface IUserRepository extends CrudRepository<User, Long> {
     User findByUsername(String username);
 }
